+++ RFKitDefaultCategories.h
@@ -27,6 +27,7 @@
 #import "UIDevice+RFKit.h"
 #import "UIImage+RFKit.h"
 #import "UIStoryboard+RFKit.h"
+//#import "UITableView+RFKit.h"
 #import "UIView+RFKit.h"
 #import "UIViewController+RFKit.h"
 //#import "UIWebView+RFKit.h"