+++ RFPerformance.h
@@ -13,7 +13,6 @@
 #import "RFRuntime.h"
 
 
-
 #ifndef RFAlloctionLog
 #   define RFAlloctionLog \
     + (id)alloc {\