+++ category/UIKit/UIDevice+RFKit.h
@@ -10,25 +10,33 @@
  */
 
 #import "RFRuntime.h"
+#import <Availability.h>
 
 @interface UIDevice (RFKit)
 
 /// If the device is iPad.
 /// Detect using userInterfaceIdiom.
-+ (BOOL)isPad;
+- (BOOL)isPad;
 
 
 /// Returns if the device has a retina screen.
 /// Detect using UIScreen's scale property.
-+ (BOOL)isRetinaDisplay;
+- (BOOL)isRetinaDisplay;
 
 
 /// Get device's mac address
 //! via: http://iphonedevelopertips.com/device/determine-mac-address.html
-+ (NSString *)macAddress;
+- (NSString *)macAddress;
 
 
-+ (long long)fileSystemFreeSize;
-+ (long long)fileSystemSize;
+- (long long)fileSystemFreeSize;
+- (long long)fileSystemSize;
+
+#pragma mark - DEPRECATED class methods
++ (BOOL)isPad DEPRECATED_ATTRIBUTE;
++ (BOOL)isRetinaDisplay DEPRECATED_ATTRIBUTE;
++ (NSString *)macAddress DEPRECATED_ATTRIBUTE;
++ (long long)fileSystemFreeSize DEPRECATED_ATTRIBUTE;
++ (long long)fileSystemSize DEPRECATED_ATTRIBUTE;
 
 @end