+++ category/UIKit/UIView+RFAnimate.h
@@ -1,0 +1,12 @@
+
+/// In experimental stages
+
+#import "UIView+RFKit.h"
+
+@interface UIView (RFAnimate)
+@property (nonatomic) CGFloat x;
+@property (nonatomic) CGFloat y;
+@property (nonatomic) CGFloat width;
+@property (nonatomic) CGFloat height;
+
+@end