+++ ChangeLog.md
@@ -1,6 +1,17 @@
 RFKit Change Log
 ================
 
+1.2
+-----
+* Receive RFFeatureSupport from RFUI/Core. Imported by RFRuntime default.
+* Class methods in UIDevice category was deprecated, new instance methods added.
+* UIView has a powerful method - isVisible.
+* Remove CGRectIsOutOfRect().
+* Add UIView+RFAnimate category.
+* NSJSONSerialization category add a parameter simplify method
+ JSONObjectWithString:.
+* Add method to deselect all selected rows for UITableView.
+
 1.1
 -----
 * Hello 2013~